+++ Dyalect/Runtime/Types/Functions/DyBinaryFunction.cs
@@ -12,7 +12,7 @@ internal sealed class DyBinaryFunction : DyForeignFunction
     private DyBinaryFunction(string name, Func<ExecutionContext, DyObject, DyObject, DyObject> fun, Par[] pars)
         : base(name, pars, -1) => this.fun = fun;
 
-    internal override DyObject InternalCall(ExecutionContext ctx, DyObject[] args) => fun(ctx, Self!, args[0]);
+    internal override DyObject CallWithMemoryLayout(ExecutionContext ctx, DyObject[] args) => fun(ctx, Self!, args[0]);
 
     protected override DyFunction Clone(ExecutionContext ctx) => new DyBinaryFunction(FunctionName, fun, Parameters);