+++ Dyalect/Runtime/Types/Functions/DyFunction.cs
@@ -1,7 +1,6 @@
 ﻿using Dyalect.Compiler;
 using Dyalect.Debug;
 using Dyalect.Parser;
-using System;
 using System.Runtime.CompilerServices;
 using System.Text;
 namespace Dyalect.Runtime.Types;
@@ -32,29 +31,37 @@ public abstract class DyFunction : DyObject
 
     public override object ToObject() => (Func<ExecutionContext, DyObject[], DyObject>)Call;
 
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    internal DyObject PrepareFunction(ExecutionContext ctx, DyObject arg)
+    {
+        var func = BindToInstance(ctx, arg);
+
+        if (func.IsExternal)
+            return ctx.NotImplemented(func.FunctionName);
+
+        ctx.EtaFunction = func;
+        ctx.Error = DyVariant.Eta;
+        return Nil;
+    }
+
     internal abstract DyFunction BindToInstance(ExecutionContext ctx, DyObject arg);
 
     internal virtual DyObject BindOrRun(ExecutionContext ctx, DyObject arg) => BindToInstance(ctx, arg);
     
-    internal abstract DyObject InternalCall(ExecutionContext ctx, DyObject[] args);
-
-    internal abstract DyObject InternalCall(ExecutionContext ctx);
+    internal abstract DyObject CallWithMemoryLayout(ExecutionContext ctx, DyObject[] args);
 
     public DyObject Call(ExecutionContext ctx, params DyObject[] args)
     {
-        var newArgs = PrepareArguments(ctx, args);
+        var newArgs = PrepareMemoryLayout(ctx, args);
 
         if (ctx.HasErrors)
             return Nil;
 
-        return InternalCall(ctx, newArgs);
+        return CallWithMemoryLayout(ctx, newArgs);
     }
 
-    protected DyObject[] PrepareArguments(ExecutionContext ctx, DyObject[] args)
+    protected DyObject[] PrepareMemoryLayout(ExecutionContext ctx, DyObject[] args)
     {
-        if (Parameters.Length == 0 && args.Length == 0)
-            return args;
-
         if (args.Length > Parameters.Length)
         {
             ctx.TooManyArguments(FunctionName, Parameters.Length, args.Length);
@@ -62,12 +69,15 @@ public abstract class DyFunction : DyObject
         }
 
         DyObject[] newLocals;
+        var needDefaults = false;
+        var memorySize = GetMemoryCells(ctx);
 
-        if (args.Length == Parameters.Length)
+        if (args.Length == memorySize)
             newLocals = args;
         else
         {
-            newLocals = new DyObject[Parameters.Length];
+            needDefaults = true;
+            newLocals = new DyObject[memorySize];
             if (args.Length > 0)
                 Array.Copy(args, newLocals, args.Length);
         }
@@ -84,14 +94,15 @@ public abstract class DyFunction : DyObject
                 newLocals[VarArgIndex] = new DyTuple(arr.UnsafeAccessValues(), arr.Count);
             }
             else if (o.TypeId != Dy.Tuple)
-                newLocals[VarArgIndex] = new DyTuple(new DyObject[] { o } );
+                newLocals[VarArgIndex] = new DyTuple(new[] { o });
         }
 
-        DyMachine.FillDefaults(newLocals, this, ctx);
+        if (needDefaults)
+            DyMachine.FillDefaults(newLocals, this, ctx);
+
         return newLocals;
     }
 
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     internal int GetParameterIndex(string name)
     {
         for (var i = 0; i < Parameters.Length; i++)
@@ -148,7 +159,7 @@ public abstract class DyFunction : DyObject
     public static bool IsSameInstance(DyFunction first, DyFunction second) =>
         ReferenceEquals(first.Self, second.Self) || (first.Self is not null && first.Self.Equals(second.Self));
 
-    internal virtual MemoryLayout? GetLayout(ExecutionContext ctx) => null;
+    internal abstract int GetMemoryCells(ExecutionContext ctx);
 
     internal abstract DyObject[] CreateLocals(ExecutionContext ctx);