+++ Dyalect/Runtime/Types/Functions/DyVariantConstructor.cs
@@ -10,7 +10,7 @@ internal sealed class DyVariantConstructor : DyForeignFunction
     public DyVariantConstructor(string name, Func<ExecutionContext, DyTuple, DyObject> fun, Par par)
         : base(name, new[] { par }) => (this.fun, Attr) = (fun, FunAttr.Vari);
 
-    internal override DyObject InternalCall(ExecutionContext ctx, DyObject[] args) => fun(ctx, (DyTuple)args[0]);
+    internal override DyObject CallWithMemoryLayout(ExecutionContext ctx, DyObject[] args) => fun(ctx, (DyTuple)args[0]);
 
     protected override DyFunction Clone(ExecutionContext ctx) => this;